bug: Fixed delete_host when host is not in a cluster.
delete_host was allowing the remote call for cluster membership to raise
a RemoteProcedureCallException. This exception was being caught and
handled returning a 404. However, when a host has no cluster membership
the host was being deleted yet a 404 was returned to the client. This
adds a narrow try/except around the cluster membership logic and handles
the exception properly.

This issue was found during e2e testing.